Finishing Schematic, Starting with Layout
Finishing and cleaning up the schematic, but also replacing some components and optimizing some footprints, for example the ATmega.
Starting with the PCB Layout - placing the components, wiring up the power circuit and the power cuircuit.
Adding an excel file, which shows the data of the chosen values of the current shunt resistors.